requirements.txt
ipython==8.11.0
numpy==1.24.2
omegaconf==2.3.0
onedrivedownloader==1.1.3
Pillow==9.4.0
pytorch_lightning==2.0.0
PyYAML==6.0
scikit_learn==1.2.2
thop==0.1.1.post2209072238
torch==2.0.0
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.15.1
tqdm==4.65.0
wandb
click==8.1.3
gdown==4.7.1